Fix stream reads on the server
There was an issue where the tcp stream was outputting
the previous write plus the new write line rather than
line by line. This was actually caused on the kvs-client.

The problem was caused by when the loop started. The string
buffer was instantiated outside the loop, so whenever the
client stdin appended to the buffer, it never kept appending
all input to the same buffer stream that was written to the
TcpStream. To fix, I moved the buffer instantiation into the
loop so that whenever a new write occurs it writes into a
fresh buffer. No previous state is maintained.